OwnTracks 6.2 RC
closes #32
closes #30
closes #25
closes #24

+++ OwnTracks/OwnTracks/Connection.m
@@ -104,7 +104,7 @@
         willRetainFlag != self.willRetainFlag ||
         ![clientId isEqualToString:self.clientId]) {
         self.host = host;
-        self.port = port;
+        self.port = (int)port;
         self.tls = tls;
         self.keepalive = keepalive;
         self.clean = clean;
